+++ docs/conf.py
@@ -18,6 +18,8 @@ extensions = [
     "sphinxcontrib.log_cabinet",
     "sphinx_issues",
 ]
+autoclass_content = "both"
+autodoc_typehints = "description"
 intersphinx_mapping = {"python": ("https://docs.python.org/3/", None)}
 issues_github_path = "pallets/itsdangerous"