+++ src/itsdangerous/__init__.py
@@ -19,4 +19,4 @@ from .timed import TimestampSigner
 from .url_safe import URLSafeSerializer
 from .url_safe import URLSafeTimedSerializer
 
-__version__ = "2.0.0rc2"
+__version__ = "2.0.0"