+++ CHANGES.rst
@@ -1,3 +1,9 @@
+Version 2.0.1
+-------------
+
+Unreleased
+
+
 Version 2.0.0
 -------------